+++ OneCoin/Server/src/test/java/OneCoin/Server/order/RedisConnectTest.java
@@ -1,4 +1,4 @@
-package OneCoin.Server.redis;
+package OneCoin.Server.order;
 
 import OneCoin.Server.helper.StubData;
 import OneCoin.Server.order.entity.RedisOrder;
