+++ OneCoin/Server/src/test/java/OneCoin/Server/order/controller/OrderControllerValidTest.java
@@ -1,4 +1,4 @@
-package OneCoin.Server.redis.redisController;
+package OneCoin.Server.order.controller;
 
 import OneCoin.Server.helper.StubData;
 import OneCoin.Server.order.dto.RedisOrderDto;
